test developer access control for item combining
